requirements.txt
coverage==3.6
Django==1.11.7
djangorestframework==3.7.3
ipdb==0.10.3
ipython==6.2.1
psycopg2==2.7.3.2
requests==2.18.4
